refactor: move main.go into ./cmd/chissoku/
move core source code into internal.
