--- version.go
@@ -1,10 +1,0 @@
-package main
-
-// ProgramName - program name
-const ProgramName = `chissoku`
-
-// Description - program description
-const Description = `A CO2 sensor reader`
-
-// Version - program versoin
-const Version = "2.1.1" // x-release-please-version
